Remove laziness around insert before lookup
Previous numbers were skewed for lookup because the insert was lazy and was
evaluated during the measurement. This change makes the insert strict so that
insert and lookup can be evaluated independently.

This change also copies the ByteStrings that will be inserted. This should
more accurately reflect reality, where lookup would likely not be done with
the same ByteString that was used during the insert.