Improved lookup performance
This change brings in some alterations to the current algorithm which greatly
reduce the memory footprint and speed up the lookup operation.

Instead of creating a tree of incremental prefixes, we create a tree with the
full prefix in each node. During lookup we also keep the key instead of
continually stripping the prefix as we traverse the tree.  In addition, we now
use an optimistic strategy when doing a lookup traversal. We only check the key
at branch points and only do a full comparison of the key if we end up at a
node with a prefix of the same length.

There is most likely an opportunity to tune the insertion code still. We are
not optimistic in our traversal because it would potentially require a fair
amount of backtracking. We may be able to get some performance improvement by
comparing only the relevant bytes (since the last branch point) during our
traversal.

As an additional benefit, this new approach has yielded slightly cleaner code
and there may be opportunities for further cleanup.